+++ src/opengl/source/viewport.cpp
@@ -55,7 +55,7 @@ struct ViewportPrivate
 
 	float waterTime;
 
-	int binaryStrength;
+	float binaryStrength;
 
 	Vec2 zoom;
 
@@ -72,7 +72,7 @@ struct ViewportPrivate
 		  rgbOffsety(Vec4(0, 0, 0, 0)),
 		  cubicTime(0.0),
 		  waterTime(0.0),
-		  binaryStrength(0),
+		  binaryStrength(0.0),
 		  zoom(Vec2(1.0, 1.0))
 	{
 		rect->set(x, y, width, height);
@@ -113,7 +113,7 @@ struct ViewportPrivate
 	bool needsEffectRender(bool flashing)
 	{
 		bool rectEffective = !rect->isEmpty();
-		bool colorToneEffective = color->hasEffect() || tone->hasEffect() || flashing || scanned || rgbOffsetx.xyzNotNull() || rgbOffsety.xyzNotNull() || zoom.x != 1 || zoom.y != 1 || cubicTime != 0.0 || waterTime != 0.0 || binaryStrength != 0;
+		bool colorToneEffective = color->hasEffect() || tone->hasEffect() || flashing || scanned || rgbOffsetx.xyzNotNull() || rgbOffsety.xyzNotNull() || zoom.x != 1 || zoom.y != 1 || cubicTime != 0.0 || waterTime != 0.0 || binaryStrength != 0.0;
 
 		return (rectEffective && colorToneEffective && isOnScreen);
 	}
@@ -172,7 +172,7 @@ DEF_ATTR_SIMPLE(Viewport, Color, Color&, *p->color)
 DEF_ATTR_SIMPLE(Viewport, Tone,  Tone&,  *p->tone)
 DEF_ATTR_SIMPLE(Viewport, Scanned, bool, p->scanned)
 DEF_ATTR_SIMPLE(Viewport, CubicTime, float, p->cubicTime)
-DEF_ATTR_SIMPLE(Viewport, BinaryStrength, int, p->binaryStrength)
+DEF_ATTR_SIMPLE(Viewport, BinaryStrength, float, p->binaryStrength)
 DEF_ATTR_SIMPLE(Viewport, WaterTime, float, p->waterTime)
 DEF_ATTR_SIMPLE(Viewport, RGBOffsetx, Vec4, p->rgbOffsetx)
 DEF_ATTR_SIMPLE(Viewport, RGBOffsety, Vec4, p->rgbOffsety)
